src/third_party/blink/web_tests/webgpu/canvas_context.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
let width = 10;
let height = 10;

test(function() {
  let canvas = document.createElement("canvas");
  canvas.width = width;
  canvas.height = height;

  let ctx = canvas.getContext('gpupresent');
  assert_true(ctx instanceof GPUCanvasContext);
}, "Test that canvas.getContext('gpupresent') returns an instance of GPUCanvasContext");

</script>
